Update README.md
In this project we have performed SQL queries in SSMS to understand our data better. The formation of queries is from basic to some advanced concepts such as cte, temptable, views. After  executing these queries we get a good understanding of the data and can use the insights for further visualization.